requirements.txt
fastapi==0.115.4
uvicorn==0.32.0
requests==2.32.3
pandas==2.2.3
pyarrow==18.0.0
pytest==8.3.3
httpx==0.27.2
email-validator>=2.2.0
pydantic>=2.0
pluggy>=1.5
apache-airflow[cncf.kubernetes,docker,http]==2.10.3+astro.1